requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy<2.0.0 # Pinning to a version below 2.0
Pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
tqdm>=4.60.0
torch==2.0.0 # Keeping this pinned for CoreMLTools compatibility
torchvision==0.19.0 # Keeping this compatible with torch 2.4.0
coremltools==6.3     # Install the latest version for compatibility, or specify a version if needed